Separate debug, &_packet, remove event list
